build validation + build audit functionality.

+++ src/main/java/com/broadcom/nbiapps/dao/SiloNameDAO.java
@@ -1,9 +1,11 @@
 package com.broadcom.nbiapps.dao;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
 
 import com.broadcom.nbiapps.entities.SiloName;
 
-public interface SiloNameDAO extends CrudRepository<SiloName, Long> {
+public interface SiloNameDAO extends CrudRepository<SiloName, BigInteger> {
     SiloName findBySiloName(String siloName);
 }
